+++ libm/mathf/internal/trigf.c
@@ -10,17 +10,17 @@
    integers, not 24 bit integers.  113 bit precision is not supported.  */
 
 static const float PIo2[] = {
-    1.5703125000e+00, /* 0x3fc90000 */
-    4.5776367188e-04, /* 0x39f00000 */
-    2.5987625122e-05, /* 0x37da0000 */
-    7.5437128544e-08, /* 0x33a20000 */
-    6.0026650317e-11, /* 0x2e840000 */
-    7.3896444519e-13, /* 0x2b500000 */
-    5.3845816694e-15, /* 0x27c20000 */
-    5.6378512969e-18, /* 0x22d00000 */
-    8.3009228831e-20, /* 0x1fc40000 */
-    3.2756352257e-22, /* 0x1bc60000 */
-    6.3331015649e-25, /* 0x17440000 */
+    1.5703125000e+00f, /* 0x3fc90000 */
+    4.5776367188e-04f, /* 0x39f00000 */
+    2.5987625122e-05f, /* 0x37da0000 */
+    7.5437128544e-08f, /* 0x33a20000 */
+    6.0026650317e-11f, /* 0x2e840000 */
+    7.3896444519e-13f, /* 0x2b500000 */
+    5.3845816694e-15f, /* 0x27c20000 */
+    5.6378512969e-18f, /* 0x22d00000 */
+    8.3009228831e-20f, /* 0x1fc40000 */
+    3.2756352257e-22f, /* 0x1bc60000 */
+    6.3331015649e-25f, /* 0x17440000 */
 };
 
 /*
@@ -52,10 +52,10 @@ static const int32_t ipio2[] = {
 };
 
 static const float
-zero   = 0.0,
-one    = 1.0,
-two8   =  2.5600000000e+02, /* 0x43800000 */
-twon8  =  3.9062500000e-03; /* 0x3b800000 */
+zero   = 0.0f,
+one    = 1.0f,
+two8   =  2.5600000000e+02f, /* 0x43800000 */
+twon8  =  3.9062500000e-03f; /* 0x3b800000 */
 
 static inline int __rem_pio2f_internal(float *x, float *y, int e0, int nx)
 {
@@ -97,7 +97,7 @@ static inline int __rem_pio2f_internal(float *x, float *y, int e0, int nx)
 
     /* compute q[0],q[1],...q[jk] */
     for (i = 0; i <= jk; i++) {
-        for (j = 0, fw = 0.0; j <= jx; j++) {
+        for (j = 0, fw = 0.0f; j <= jx; j++) {
             fw += x[j] * f[jx + i - j];
         }
 
@@ -117,7 +117,7 @@ static inline int __rem_pio2f_internal(float *x, float *y, int e0, int nx)
 
         /* compute n */
         z  = scalbnf(z, (int32_t)q0);   /* actual value of z */
-        z -= (float)8.0 * floorf(z * (float)0.125); /* trim off integer >= 8 */
+        z -= 8.0f * floorf(z * 0.125f); /* trim off integer >= 8 */
         n  = (int32_t) z;
         z -= (float)n;
         ih = 0;
@@ -129,7 +129,7 @@ static inline int __rem_pio2f_internal(float *x, float *y, int e0, int nx)
             ih = iq[jz - 1] >> (7 - q0);
         } else if (q0 == 0) {
             ih = iq[jz - 1] >> 7;
-        } else if (z >= (float)0.5) {
+        } else if (z >= 0.5f) {
             ih = 2;
         } else {
             /* No action required */
@@ -189,7 +189,7 @@ static inline int __rem_pio2f_internal(float *x, float *y, int e0, int nx)
                 for (i = jz + 1; i <= jz + k; i++) { /* add q[jz+1] to q[jz+k] */
                     f[jx + i] = (float) ipio2[jv + i];
 
-                    for (j = 0, fw = 0.0; j <= jx; j++) {
+                    for (j = 0, fw = 0.0f; j <= jx; j++) {
                         fw += x[j] * f[jx + i - j];
                     }
 
@@ -203,7 +203,7 @@ static inline int __rem_pio2f_internal(float *x, float *y, int e0, int nx)
     } while (recompute);
 
     /* chop off zero terms */
-    if (z == (float)0.0) {
+    if (z == 0.0f) {
         q0 -= 8;
 
         for (jz -= 1; jz>=0; --jz) {
@@ -236,7 +236,7 @@ static inline int __rem_pio2f_internal(float *x, float *y, int e0, int nx)
 
     /* compute PIo2[0,...,jp]*q[jz,...,0] */
     for (i = jz; i >= 0; i--) {
-        for (fw = 0.0, k = 0; k <= jp && k <= jz - i; k++) {
+        for (fw = 0.0f, k = 0; k <= jp && k <= jz - i; k++) {
             fw += PIo2[k] * q[i + k];
         }
 
@@ -244,7 +244,7 @@ static inline int __rem_pio2f_internal(float *x, float *y, int e0, int nx)
     }
 
     /* compress fq[] into y[] */
-    fw = 0.0;
+    fw = 0.0f;
 
     for (i = jz; i >= 0; i--) {
         fw += fq[i];
@@ -279,14 +279,14 @@ static inline int __rem_pio2f_internal(float *x, float *y, int e0, int nx)
  */
 
 static const float
-half =  5.0000000000e-01, /* 0x3f000000 */
-invpio2 =  6.3661980629e-01, /* 0x3f22f984 */
-pio2_1  =  1.5707855225e+00, /* 0x3fc90f80 */
-pio2_1t =  1.0804334124e-05, /* 0x37354443 */
-pio2_2  =  1.0804273188e-05, /* 0x37354400 */
-pio2_2t =  6.0770999344e-11, /* 0x2e85a308 */
-pio2_3  =  6.0770943833e-11, /* 0x2e85a300 */
-pio2_3t =  6.1232342629e-17; /* 0x248d3132 */
+half    =  5.0000000000e-01f, /* 0x3f000000 */
+invpio2 =  6.3661980629e-01f, /* 0x3f22f984 */
+pio2_1  =  1.5707855225e+00f, /* 0x3fc90f80 */
+pio2_1t =  1.0804334124e-05f, /* 0x37354443 */
+pio2_2  =  1.0804273188e-05f, /* 0x37354400 */
+pio2_2t =  6.0770999344e-11f, /* 0x2e85a308 */
+pio2_3  =  6.0770943833e-11f, /* 0x2e85a300 */
+pio2_3t =  6.1232342629e-17f; /* 0x248d3132 */
 
 int32_t __rem_pio2f(float x, float *y)
 {
